# Add a perf scenario for DotNetDispatcher