--- a/src/JSInterop/Microsoft.JSInterop/perf/AssemblyInfo.cs
+++ b/src/JSInterop/Microsoft.JSInterop/perf/AssemblyInfo.cs
@@ -1,0 +1,1 @@
+[assembly: BenchmarkDotNet.Attributes.AspNetCoreBenchmark]