I tried playing around with using `ObjectResultExecutor` to avoid the reflection invoke in `DotNetDispatcher`. It does seem like a good change. We could consider a strategy where it does a reflection invoke the first few times around, but produces a compiled delegate if it's more frequently called. I recall DI has (❓) a pattern like this. 

# Vanilla benchmark

 Method |     Mean |     Error |    StdDev |      Op/s |  Gen 0 | Allocated |
------- |---------:|----------:|----------:|----------:|-------:|----------:|
 Baseline | 1.220 us | 0.0103 us | 0.0080 us | 820,004.1 | 0.0057 |     240 B |
 Invoke | 1.002 us | 0.0064 us | 0.0060 us | 998,215.6 | 0.0038 |     200 B |
 
 
# 5000 operations per invoke
 Method |     Mean |    Error |   StdDev |        Op/s |   Gen 0 | Allocated |
------- |---------:|---------:|---------:|------------:|--------:|----------:|
 Baseline | 1.197 us | 0.0099 us | 0.0093 us | 835,661.3 | 31.2500 |   1.14 MB | 
 Invoke | 964.1 ns | 11.89 ns | 11.12 ns | 1,037,280.3 | 23.4375 | 976.56 KB |