@pranavkm 2019-09-17
Fair enough. Let's get the infrastructure for measuring the WASM size impact before we try this out.